backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==4.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.2
cssselect2==0.8.0
daphne==4.2.0
dj-database-url==3.0.1
Django==4.2.20
django-admin-tabler==1.0.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==25.0
paho-mqtt==2.1.0
paramiko==3.5.1
pillow==11.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
requests==2.32.3
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zope.interface==7.2
zopfli==0.2.3.post1
